This commit introduces a new ServletContextListener, ApplicationInitializer, which offloads parts of the responsibilities from EntryServlet. The EntryServlet was both serving as Jersey entry point servlet, but also as handling application initialization. This is now done in the listener. The ConfigurationFactory is also throwing an exception when there is no configuration file, instead of printing to stderr and later throwing a NullPointerException.
With these changes, it is easy to identify when the application fails to initialize.
Which is useful when a WAR application is deployed to a platform like AWS Elastic
Beanstalk.

Simplify
