+++ src/main/webapp/WEB-INF/web.xml
@@ -11,6 +11,11 @@
         <param-value>edu.illinois.library.cantaloupe.RestletApplication</param-value>
     </context-param>
 
+    <listener>
+        <display-name>ApplicationInitializer</display-name>
+        <listener-class>edu.illinois.library.cantaloupe.ApplicationInitializer</listener-class>
+    </listener>
+
     <servlet>
         <servlet-name>ServerServlet</servlet-name>
         <servlet-class>edu.illinois.library.cantaloupe.EntryServlet</servlet-class>
@@ -18,7 +23,6 @@
             <param-name>org.restlet.clients</param-name>
             <param-value>CLAP HTTP HTTPS</param-value>
         </init-param>
-        <load-on-startup>0</load-on-startup>
     </servlet>
 
     <servlet-mapping>
